+++ lib/ComponentManager/PackageSource/ZipPackageSource.php
@@ -104,7 +104,7 @@ class ZipPackageSource extends AbstractPackageSource
     /**
      * @inheritdoc PackageSource
      */
-    public function obtainPackage($tempDirectory,
+    public function obtainPackage($tempDirectory, $timeout,
                                   ResolvedComponentVersion $resolvedComponentVersion,
                                   Filesystem $filesystem,
                                   LoggerInterface $logger) {